requirements.txt
Flask==2.0.1
python-dotenv==0.19.1
google-auth==1.35.0
google-auth-oauthlib==0.4.6
openai==1.109.1
pandas==2.3.2
numpy==2.3.3
streamlit==1.50.0
PyPDF2==3.0.1
requests==2.32.5
python-docx==1.2.0
python-pptx==1.0.2
XlsxWriter==3.2.9